[Fix] Movement, Shooting (Enemy, Player)
Additionally changed movement to take into an account DeltaTime so it wont be faster on certain platforms and slower on others and stable same instead.

Shooting for both enemy and player did some refactoring so that the bullets are facing the correct way around.

Movement refactored into a smaller collection easier to understand.

Fixed the Spaceman Holding his hands and added the pistol to it.